services/scraper/local_agent/requirements.txt
# Local-agent-only deps (run on user's PC, NOT installed on Railway)
playwright>=1.45
websockets>=12.0
httpx>=0.27
pydantic>=2.6
python-dotenv>=1.0

# Desktop control (browser-only users can skip these — see local_agent/README.md)
pyautogui>=0.9.54
mss>=9.0
pillow>=10.0
pynput>=1.7
openpyxl>=3.1
